Jarvis_Level_1/requirements.txt
openai
python-dotenv
speechrecognition
requests
edge-tts
playsound==1.2.2
pyaudio
